+++ Source/Aul2AudioViewRender.pas
@@ -22,6 +22,7 @@ uses
   Aul2AudioViewRenderMirrorBars,
   Aul2AudioViewRenderPixelWave,
   Aul2AudioViewRenderPulseWave,
+  Aul2AudioViewRenderVectorscope,
   Aul2AudioViewRenderWaveLine;
 
 const
@@ -65,6 +66,7 @@ begin
     VIEW_TYPE_PULSE_WAVE: DrawPulseWave(Buffer, Width, Height, Settings, CurrentFrame);
     VIEW_TYPE_CIRCULAR_SPECTRUM: DrawCircularSpectrum(Buffer, Width, Height, Settings, CurrentFrame);
     VIEW_TYPE_MIRROR_BARS: DrawMirrorBars(Buffer, Width, Height, Settings, CurrentFrame);
+    VIEW_TYPE_VECTORSCOPE: DrawVectorscope(Buffer, Width, Height, Settings, CurrentFrame);
   else
     DrawEqualizerBars(Buffer, Width, Height, Settings, CurrentFrame);
   end;